+++ CHANGELOG.md
@@ -5,7 +5,7 @@ _Breaking changes_
 - Revised and improved the navigation.
 - Revised and improved the Queue.
 - Improved FxResponsive Widget
-- changed hide() to hide
+- changed hide() to accept condition and removed hideIf and showIf
 
 _Added_
 
@@ -13,6 +13,7 @@ _Added_
 - Added shadow mixin
 - Added FxButton widget
 - Added FxTextFormField Widget
+- pick method added into map extension
 
 ## 3.6.4